refactor(ui): extract fixBrokenIcons into ui/icons.ts
Move the 116-line SVG-icon repaint helper out of main.ts into
src/ui/icons.ts. Pure DOM with static markup, no application state.
main.ts: 2050 -> 1934 lines.

Verified: typecheck/lint/build clean, e2e 6 pass / 1 skip.
